refactor auth flow to pass token to server instead, better for mfa later
